refactor(spec-546): shrink the shared interface — 59 exports to 24 (#671)
An `export` is a promise: this is a supported seam, build on it. An agent
picking a helper out of agent/handlers/shared.ts saw 59 candidates and could
not tell resolveRefArg (11 consumers, a real seam) from agoLabel or
statusNextAction (zero). Measured against std-51: 18 exports had no consumer
outside the file at all, and 5 more were referenced only by source-text scans,
never imported.

- Dropped `export` from 23 symbols. No logic, no reordering, no renames.
- Moved 12 single-consumer symbols to the sibling handler that was already
  their only caller, private on arrival: comments.ts x3, tasks.ts x3,
  standards.ts x3, plus acs.ts, roles.ts, lifecycle.ts.
- Deleted parseTypeFilter as dead code — zero consumers, and
  routes/comments.ts:84 carries its own private implementation of the name.

shared.ts: 1810 -> 1531 lines, 59 -> 24 exports. 8 files, all under
agent/handlers/ bar a comment-only tool-specs.ts edit; none of the 56
importers changed an import specifier.

Five symbols that look like they belong with their consumer deliberately stay
(dec-2, ac-2): renderFooterSignal, craftActivityBlock, composeStatusOverview,
StatusFacts and relatedIssuesNudge each have one consumer too, but it is the
tool-specs.ts façade. Moving them there takes them outside the only directory
guidance-authoring-confined scans, and its `while (idx !== -1)` loop then
never runs its body: the test would pass while asserting nothing and the "one
author of footer prose" invariant would go unguarded, in green.

handoffInterpolationContext does NOT move — shared.ts calls it itself, giving
it two consumers — so the count lands at 24, not the 23 planned.

No test added, deliberately: a test asserting an export count would gate
interface size, which std-51 cl-34 forbids and this Spec lists as out of scope.

Verified: make typecheck exit 0; real .d.ts emission exit 0 separately, since
--noEmit skips declaration emit and tsconfig sets declaration:true; full
server suite 707 files / 6552 tests passed, identical counts before and after;
the seven source-scanning guards run explicitly, 45 tests passed, unmodified.

Spec: https://memex.ai/mindset-prod/memex-building-itself/specs/spec-546
